stop test execution

+++ app/Observers/GroupUserObserver.php
@@ -15,7 +15,7 @@ class GroupUserObserver
         if ($groupUser->group->type === GroupTypeEnum::Department) {
             CheckStaffGroupMembershipJob::dispatch($groupUser->user);
         }
-        if ($groupUser->group->nextcloud_folder_name) {
+        if ($groupUser->group->nextcloud_folder_name && !app()->runningUnitTests()) {
             NextcloudService::addUserToGroup($groupUser->group, $groupUser->user);
             $allowAclManagement = in_array($groupUser->level, [GroupUserLevel::Admin, GroupUserLevel::Owner]);
             if ($allowAclManagement) {
@@ -26,7 +26,7 @@ class GroupUserObserver
 
     public function updated(GroupUser $groupUser): void
     {
-        if ($groupUser->group->nextcloud_folder_name) {
+        if ($groupUser->group->nextcloud_folder_nam && !app()->runningUnitTests()) {
             if ($groupUser->isDirty('level')) {
                 $allowAclManagement = in_array($groupUser->level, [GroupUserLevel::Admin, GroupUserLevel::Owner]);
                 NextcloudService::setManageAcl($groupUser->group, $groupUser->user, $allowAclManagement);
@@ -39,7 +39,7 @@ class GroupUserObserver
         if ($groupUser->group->type === GroupTypeEnum::Department) {
             CheckStaffGroupMembershipJob::dispatch($groupUser->user);
         }
-        if ($groupUser->group->nextcloud_folder_name) {
+        if ($groupUser->group->nextcloud_folder_name && !app()->runningUnitTests()) {
             NextcloudService::removeUserFromGroup($groupUser->group, $groupUser->user);
             NextcloudService::setManageAcl($groupUser->group, $groupUser->user, false);
         }
